Move time and status display to top
